added some screen cracks to notify of low health, when player dies the map generates and enemies are reset, craters now reset once map generates or game reset is triggered, enemies reset their health on activation, instead of triggering all enemies dead i trigger map generation for the time being

+++ Components/ObjectTypes/DamageData.cs
@@ -2,13 +2,13 @@ using Godot;
 
 public class DamageData
 {
-    public int Damage { get; }
+    public float Damage { get; }
     public float Knockback { get; }
     public Node Source { get; }
     public string WeaponName { get; }
     public string DamageType {get;}
 
-    public DamageData(int damage, float knockback, string name, string type)
+    public DamageData(float damage, float knockback, string name, string type)
     {
         Damage = damage;
         Knockback = knockback;
